Add a very basic math list -> horizontal list converter
This starts the framework for converting a math list into a horizontal
list, as detailed in Appendix G of the TeXbook. After a few different
iterations of how to structure this, I've decided on two loops each
with a big `match` to process the elements. I will continue to build
on this to continue handling all of the rules in the appendix.

One question I've been trying to understand is why TeX doesn't show
nested boxes when printing out math boxes. For instance,
```
\setbox0=\hbox{$ab$}\showbox0
> \box0=
\hbox(6.94444+0.0)x9.57755
.\mathon
.\teni a
.\teni b
.\mathoff
```
I would expect to see an extra level of box around `\teni a` and
`\teni b` based on rule 17, but maybe it's not clear what it means by
"Convert the symbol to a character box". In any case, I don't believe
that wrapping these individual characters in boxes will cause any
problems for now, until I discover otherwise.