Add inter-atom spacing in math lists
This adds to the math list -> horizontal list conversion by
determining how much space to add between each element based on the
atom kinds. Right now I'm just using a constant amount of space
defined in `pt` instead of `mu`, but it shows that the process is
working.

The test that I wrote for this is not technically correct; the fonts
are technically wrong, and in reality some of the characters should be
shifted up/down based on their kinds, but it makes it easy to ensure
that I'm handling the different pairs correctly for now.