Parse math style changes, keep track of current style
This adds parsing for style change elements in math lists like
`\displaystyle`. We can keep track of the current style while
converting a math list to a horizontal list, and then use that when
determining the space between atoms.

I even had a test lying around to test something which had a style
change in it, and I added some more tests around the spacing between
atoms.