Use scaled font metrics for scaled fonts
This follows up on the previous commit where I added scaled
fonts. This adds a new `FontMetrics` primitive that loads TFM files
and returns the metrics scaled by the `scale` in the `Font`. Most of
this PR is going around replacing usages of `TFMFile` with
`FontMetrics`.

The more major change in here is making font metrics in `TeXState`
loaded on-demand, instead of pre-loading the one font that is
currently used. I used another `RefCell` to allow the font metrics to
be changed at any time, and learned about how `Ref` works so that I
can return a reference to something stored in the `RefCell` using
`Ref::map`. I previously would have probably liked to do something
similar with boxes, but ended up just making copies and/or taking the
boxes out instead of returning references.