Parse \font definitions
This adds parsing \font definitions. These aren't used for anything
yet, I will make fontdefs set the current font next.

Parsing font definitions was relatively straight forward. To see what
characters I needed to support for filenames, I ran
```
find /usr/local/texlive/2018/texmf-dist/fonts/tfm/public/ | \
  grep tfm | \
  xargs basename -s .tfm | \
  grep -v '^[a-zA-Z0-9_-]\+$'
```
and saw that it returned no results, so I think I can away with only
supporting alphanumeric and `-` and `_` characters.

The one confusing bit was this section from the TeXbook that says "TeX
takes precautions so that constructions like `\font\cs=name\cs` won’t
expand the second `\cs` until the assignments are done." It sounds
like there might be a way to indicate that the expansions of a certain
token should be ignored. We don't have a way to do that right now, so
I resorted to just setting the token equal to `\relax` which should
have a vaguely similar effect in this instance. (If someone has
reassigned `\relax`, then this will cause problems, which is why it's
not a perfect solution.) The one gotcha was that, in order to test the
`\font\cs=name\cs` construct appropriately, I need to attempt to load
an invalid font, which will end up throwing an error, so the best I
can do is check the error message that is thrown for now.