Allow changing the current font using fontdef tokens
This allows changing the current font by assigning using a fontdef
token. This kind of assignment looks a little different than most of
the other ones, because it's impossible to know whether we're looking
at an assignment without checking if a token is assigned to a fontdef
token. This is relatively straightforward; unfortunately, the way I
have it set up also means that a few extra copies of the stored font
have to be made just to check if it exists.