Pick correct fonts for math characters based on style and family
Now that I can change fonts in horizontal lists, I can update the math
converter to use the correct fonts for math characters. The actual
correct way to do this is to pull from `\textfont`, `\scriptfont`, and
`\scriptscriptfont`, but for now I'm just hardcoding the default fonts
for families 0-3.

In the process, I noticed that I had never fixed a bug from a while
ago where the assignments from inside `\hbox{}` and `\vbox{}` leaked
into the outer environment. In order to make some of the tests work, I
fixed that up.