Ignore spaces in math mode
This ignores space tokens when in math mode. I actually expected
`\par` to also be ignored in math mode, but it turns out that `\par`
is invalid in math mode. Maybe I just haven't written TeX in a while,
but I thought you could have blank lines in math mode. Apparently
not, even in display math!