Add translation of superscripts, misc requirements for superscripts
This adds support for converting math expressions with superscripts in
them (e.g. `$a^b$`) into horizontal lists. There were many steps
involved in getting this to work, like adding access to font
dimensions and adding some helper methods to mathstyles.

I also realized that the "first pass" translation that happens when
converting a math list into a horizontal list doesn't produce a normal
math list again, and in fact it was impossible to convert the result
to a math list. I made a separate, intermediate step format for
capturing this intermediate data, which also cleaned things up because
some of the assumptions that are made in the second pass can be baked
into the types.

I uncovered a problem with how I handle font metrics, where if I held
onto a font metrics reference while I was trying to load the font
metrics for a new font, this would cause a crash. I realized that I
need to hold onto font metrics references for as short as possible to
prevent this, so I added a helper function to make this easier.

The infrastructure is also now in place to convert subscripts and
superscript/subscript pairs, so I'll work on those next.