Add translation of subscripts in math mode
Following with the structure for how superscripts are handled, this
adds the translation of subscripts. I was going to also handle
superscript/subscript combinations, but noticed that there was a bug
with how single character boxes are translated, which meant that
`$a_{b^c}$` ends up getting translated incorrectly. Interestingly, it
only happens in scriptstyle prime, not plain scriptstyle, so
`$a^{b^c}$` works fine.

I added a skipped test for this and will fix this next, then move onto
combo superscript/subscripts.