Fix individual symbol nuclei being translated correctly
Previously, all atom nuclei would be wrapped in a box when they were
transformed. Looking more closely at the TeX output, not only is this
incorrect, it caused some translation errors for things like
`$a_{b^c}$` because we weren't handling the case of "If the
translation of the nucleus is a character box..." correctly.

This solves both of those problems by not wrapping single symbols in
boxes, and also makes it possible to correctly handle the
superscript/subscript case correctly.

It turns out that this affected a whole bunch of tests that were
assuming nuclei spit out hboxes, but that was relatively easily fixed.