Allow \prevdepth assignments in vertical modes
When trying to generate a duplicate version of the output of
superscript+subscript combos using `\hbox{}` and `\vbox{}` I ran into
a new issue of interline glue. Because of interline glue, it's
difficult to construt exactly the `\vbox{}` that I desire for the
tests to compare against. There is a way to disable interline glue, by
setting `\prevdepth=-1000pt`. However, I wasn't handling `\prevdepth`
assignments, so I looked into implementing those.

After reading through the TeXbook chapter 24 several times, I finally
wrapped my head around how these kinds of assignments are supposed to
work. I believe that the place I currently have my prev_depth value
stored, in the routine for parsing vertical lists, is the correct
place, because it's not supposed to be stored in the global state but
is still local to each vertical box. However, I needed to give access
to the value for assignment purposes.

I decided to just pass in a mutable reference to that specific
variable to the assignment routine and let that routine operate on
it. There are other values that are going to be operated on in similar
ways, so I put these into a struct to handle all of them. There are
few enough of them that this should work fine.