Add support for superscript/subscript combinations
This combines the code from the superscript and subscript routines to
generate combined superscripts+subscripts. The math needed to
calculate the shifts is a little more complicated, but not too
much. Also, this generates a vertical box instead of just a shifted
box.

It took me a while to get the math here figured out, mostly because
the height of the resulting boxes kept ending up different from what
TeX produces. I eventually realized that the instructions in the
TeXbook are slightly incorrect, instead of a vbox with height `h(x) +
u` and depth `d(y) + v`, a vbox with height `h(x) + u + v` and depth
`d(y)` is produced, with an added shift of `v`. This ends up looking
exactly the same, but it's easier to just make my code do the same
thing as TeX, so I went with that.

As a side note, I'm really enjoying writing the tests for these math
conversion cases. Instead of trying to describe the output using
structs, etc, just finding the equivalent representation in TeX makes
things easy to test and easy to write.