Allow converting between math atom types during the first pass conversion
Before the second pass of a math list where kerns are added between
elements in the math list, many different atom types are converted to
ords. This adds in that step for ensuring that the different types are
appropriately converted.

For now, since Vcent, Over, Under, Rad, and Acc atoms aren't handled,
the only conversion that happens is turning Bins into Ords in certain
situations. Interestingly, some of this conversion happens during the
processing of the next atom, so we need to look back in the existing
list to find the Bin atom to change.

I wish that I had a better way to test to ensure that this conversion
was happening by inspecting the actual results of the mid-conversion
list, but for now I'm just using the fact that the space between
different atom types is different to ensure that the conversions are
happening.