Raise op atoms to be centered on the baseline
This implements the rule that op atoms are shifted so that they are
centered along the axis, to match up with other symbols. I handled
this by doing the conversion when the nucleus is converted into a
horizontal list element, because it needs to be wrapped in a box in
order to be shifted.

I also learned that in the font, some operator symbols like \sum and
\int are not even close to being centered along the axis, they lie
almost entirely below the baseline. I guess when the symbols are going
to be centered for you, the font creator doesn't have to really work
on ensuring that they are centered!