Add routine for reboxing a box to a given width
This is a routine that is used in operator limits as well as
generalized fractions. It wraps the contents in `\hfil`s and puts the
result in a new horizontal box of the specified width.

In the tests, I feel like I could theoretically get away with just
comparing a reboxed `\hbox{a}` to a `\hbox to 10pt{\hfil a\hfil}`, but
something about that makes me not feel like I'm entirely covering the
essence of the reboxing, so I added in extra checks for the final
widths and that the glue set ratios were correct.