Add basic generalized fraction parsing
This adds basic support for parsing generalized fractions, which
happens using a special state within the math list parsing. This sets
the special state when a generalized fraction is encountered, and does
extra parsing at the end of the math list.

This doesn't correctly render these generalized fractions yet, that
will happen next.