Add rendering of basic generalized fractions (i.e. \atop)
This adds the ability to render generalized fractions without bar
lines or delimiters. The algorithm for this is relatively simple
compared to that of superscripts/subscripts. I didn't implement
fractions with a bar line because I don't currently support rules. I
added a helper function to eventually generate the delimiters that
might go on the left and right of the fraction, but for now it only
generates null delimiters.

Writing the tests for these was somewhat maddening, but I figured it
out in the end. There's no easy way to generate `\vbox{}`es with
specified height and depth, which is necessary in this case because
the boxes actually have a height and depth (unlike
superscripts/subscripts which were boxes with only height that had
been shifted).