Fix superscript/subscript shifts
When implementing superscripts/subscripts originally, I apparently
misread "where q and r are the values of σ_18 and σ_19 in the font
corresponding to styles C↑ and C↓" and skipped that part about the
metrics not being from font `C`.

I noticed this because superscripts and subscripts on something like
`{{a \atop b} \atop {c \atop d}}` ended up being in the wrong place
compared to TeX. I'm guessing that in most of the cases I've tested,
those metrics were getting overridden by something else, but not in
the case of larger nuclei.

Figuring out what was wrong here was unfortunately very difficult, and
took me reading through the steps many times before I noticed the
mistake. I hope that the tests are now sufficient that I can trust
these more "fundamental" operations for the future.