Turn inter-atom skips into muskips
This introduces MuDimens and MuGlue which is used instead of normal
glue in the skips between atoms. MuDimens are essentially dimensions
that are defined in terms of 1/18 of an em, which allows them to scale
propotional to the font (like how ems do), which matters in math
because they should scale with the style of the font.

I added the minimum necessary to get this working, so I added a new
MuDimen struct and MuGlue struct with utility functions for converting
them into normal Dimens and Glue. These are currently only used in
determining the skips in between atoms, though.

One fun little quirk I ran into was that the conversion from MuDimen
into Dimen actually loses some precision along the way, so if I'm
using a font where a quad (i.e. the width of an em-dash) is 10.0pt and
I create a `\mskip 18mu`, I would expect to get a skip of exactly
10pt, but I instead get 9.99976pt. I had to do some messing around to
get my implementation to achieve the same loss of precision.