Round GlueSetRatio when converting from f64 -> i32
Rounding in TeX is a little inconsistent. For instance, doing

```
\dimen0=0.000029pt
```

will set `\dimen0` equal to 2 scaled points, which is about 0.00003,
so the dimension value is clearly being rounded up. However, if I do

```
\dimen0=5sp
\divide\dimen0 by2
```

then `\dimen0` will now be equal to 2 scaled points, not 3, as it
might be if it was rounded. So clearly, when doing integer operations,
the values are floored, not rounded.

It turns out that glue set ratios are similar to dimens, in that they
are rounded to the nearest integer when they are initialized. (They
aren't editable, so the flooring don't apply to them.) In Rust,
converting from a floating point to an integer value will floor the
value, so I have to explicitly round it in the code in order to
achieve this same behavior.

I noticed this because in the final stage test, one of the elements
was exactly one sp off from the output that TeX gives. This was
because when centering an element in a generalized fraction, we
surround something with 2 `\hfil`, so the extra space that needs to be
filled in gets divided by 2, and if that space is an odd number of sp,
it should be rounded up instead of floored.

I will eventually get around to fixing the issue of Dimens needing to
be rounded as well.