Start stage 6, line breaking
I finally decided on the goal for stage 6, which is going to be line
breaking. After reading through chapter 14 of the TeXbook several
times, I am kinda convinced that the line breaking algorithm isn't
going to be super hard to implement, so I'm adding in some extra
challenges (like discretionary breaks) at the same time. Obviously I
will start with the basics first, but I don't think the other things
will add much to the complexity.

I have implemented the Knuth-Plass linebreaking algorithm before in
JavaScript for formatting text in a webpage, so I think I understand
how the algorithm is supposed to be implemented.