Update dependencies, switch to once_cell
I updated kpathsea because there was a security warning about a
dependency of the current version. While I was messing with
dependencies, I switched things over from `lazy_static!` to
`once_cell`'s `Lazy`, which is likely to be standardized so will be
better to use in the long term.