Move the glue setting utilities out of the parser module
Glue setting is a process that I am going to need for line breaking,
which I don't want to have to happen within the context of a
parser. This pulls the glue setting utilities out of the parser and
into the `boxes` module, including the utility which combines a vec of
horizontal list elems into a box.

In addition, I added an intermediate result type to the glue setting
process, which helps represent some of the soft errors that can occur
when setting glue (notably, not enough shrink and not enough
stretch). This intermediate type can be collapsed into an exact glue
set, but the intermediate reporting of errors will help with line
breaking.