Add initial line splitting algorithm
This adds an initial line splitting algorithm which works based on the
description in the TeXBook. Currently, it generates a directed graph
of all possible breaks in a paragraph, then performs a search through
the nodes to find the least number of demerits to the end node. The
current implementation is very naive and can (and will) be sped up,
but is good enough for now. (I tried making a smarter implementation
without a graph search to start with, but it ended up being too
complicated and produced incorrect results, so this is preferable).

I tested this by generating some paragraphs and feeding them into TeX
and seeing how TeX splits them, and then writing tests to make sure
that this implementation matches what TeX does. The exact number of
demerits don't match because there are certain penalties that aren't
taken into account yet, but it is exactly the same otherwise.