Clean up some warnings from test-only code
Some old code and new modules are only referenced in tests (for now),
which causes dead_code warnings. Instead of just suppressing dead_code
warnings entirely (or just leaving these warnings for now), this marks
many of those things as test-only.

There are still many dead_code warnings in the dvi module. These are
more difficult to deal with because different parts of the module are
referenced by different binaries. This causes warnings because these
binaries are compiled separately so different warnings depending on
which binary is being run. One potential solution online is to make the
dvi code a separate library which is imported by the rest of the code.
Then, it would only be compiled a single time into the library so all of
the warnings would be generated once. Maybe an improvement for the
future.