Add parsing of glue parameters
This allows parsing of glue variables for both setting them and
retrieving them when parsing glue. Parsing glue is a tiny bit
complicated because something like "-\parskip" and "-3pt plus1pt" are
parsed differently, but both start with - signs. Instead of trying to
distinguish them immediately, we parse the initial signs and multiply
whatever comes next by the value of the signs.

In the process of doing that, I noticed that the way signs are parsed
was slightly incorrect. "-   " was parsed correctly but "- - " was not.
It turns out that spaces in between negative signs are allowed. I fixed
that as well.