Use line breaking to break vertical lists
This applies the new line breaking algorithm to actually break vertical
lists into a series of horizontal boxes. The main change here is in
`parser::vertical_list::handle_enter_horizontal_mode`, which previously
simply returned a single horizontal box. In addition to applying line
breaking, this function also handles some slight modifications to the
horizontal list, like adding the `\parfillskip` before the end. Without
these modifications.

To accomplish this, I had to add `\parfillskip` to the list of glue
parameters and added its default value.

This ended up affecting many tests in other places which made
assumptions about how vertical boxes were parsed. In particular, I
increased the `\hsize` in many of the tests so that no line wrapping
would be applied. Also, while the the `\parfillskip` addition was an
invisible change in most places, it is still written to the output DVI
files which affected some of the tests there.