Actually implement the Knuth-Plass algorithm
This was a big one. When trying to deal with overfull boxes, the naive version
of my line breaking algorithm fell apart. I ended up folding and looking up some
details of how the Knuth-Plass algorithm works, which wasn't too complicated.
(It's still just a graph search, but has some smart optimizations for the
constraints that our search has). In particular, it's very clear where the
overfull box "escape hatch" is supposed to live, because the nodes of the line
breaks are visited in order. My previous algorithm was causing problems because
the line breaks were visited in a arbitrary order, so ensuring that the overfull
box that was created was truly the smallest one possible was hard(/impossible)
to do.

After getting that working, I ran into some other issues with matching TeX's
output. Namely, I found that my (what I thought wasn't naive) calculation of
badness was *very slightly* off from what TeX was producing. After many hours of
banging my head on the wall to understand how the approximation was being made,
I caved to one of my main principles (not looking at the source code of TeX!) to
see how it was done for real. Thankfully, the source has a nice comment telling
me that I should use *that specific approximation exactly*, so I don't feel too
bad about cribbing from the source.

While doing all this, I ended up building in the basics of the debug printout
that TeX generates for the paragraph algorithm. (Another thing that would have
bit me hard later if I had used my naive algorithm: the printout shows an
ordered traversal of the breaks, so I would have had to reconstruct it, instead
of just being able to print as I went!) Conveniently, I had just added integer
parameters, so I can even toggle the debugging on and off in the same way that
TeX does (instead of a compiler flag or something more sensible).

In the process of implementing this, I have seen many different avenues that I
have left still unexplored. In particular, I've noticed that handling "visually
incompatible" lines is going to be very important to getting the line breaking
correct. In fact, TeX's graph search (I think) allows multiple edges from the
same node if they produce different visual incompatibilities! One of the tests
is currently producing incorrect output because of this, so I will need to fix
that up.