Clean up `LineBreakGraph`
As the line breaking algorithm has grown, less and less responsibility has been
taken up by the `LineBreakGraph` struct and more has been taken up by the
`generate_best_list_break_option_with_params` function itself. This removes some
now-unneeded complication in how `LineBreakGraph` works, by storing the paths
through the graph as a map from point to point, instead of storing it as an
array that is tied to an array of points.

This drastically simplifies a few of the functions of `LineBreakGraph`, and
because the handling of choosing points to look at is handled elsewhere, we
don't need an ordered list of the points stored there anyway.