Add in visual incompatibility demerits
This adds in the calculation of the "visual classification" of a line's setting,
depending on how much the glue inside the line has been stretched or shrunk.
When adjacent lines have differently stretched glue, they look worse next to
each other, so we try to minimize those situations by adding extra demerits.

While doing this, I realized that demerits can actually be negative! Very fun.
(Probably more relevant when you have constructs like `\penalty-5000` but not
there yet) I cleaned up some code around there for that. At some point, I need
to test if there are numeric maxima for the different parameters involved,
because right now I'm using 64-bit values which probably are extreme for this.

I also added in the new `\adjdemerits` integer parameter, and in the process I
decided to clean up the way that variables which store to parameters are
handled. Instead of a separate enum value for each parameter in each variable,
I'm just wrapping the parameters. This will make it a lot nicer as I add more
and more parameters.