Make some small line breaking code cleanups
I noticed before when debugging that the numerical line break identifiers being
printed out when `\tracingparagraphs=1` is set had gaps in between. E.g.

```
@ via @@0 b=30 p=x d=1600
@@3: line x.1 t=1600 -> @@0
@ via @@0 b=35 p=x d=2025
@@5: line x.3 t=2025 -> @@0
@ via @@1 b=44 p=x d=3016
@@8: line x.3 t=4616 -> @@3
```

I now only add line break numbers when they are actually going to be shown.

While I was at it, I decided that the line breaking code really should be mostly
independent of the TeX state, with the parameters being stored in a separate
object. The decision of logging debugging info was breaking that and was reading
directly from state. This adds a separate new param which can be used for
turning on the debugging. (Some sort of state is still needed here, but it's
only being used to load and read font metrics.)

Finally, I like the pattern in my most recent tests where I am rendering the
same input with different parameters to see the difference that it makes. I
refactored out the input to a variable in an existing test to make it clear that
the input is actually the same thing.